data/requirements.txt
mysql_connector_python==9.0.0
pandas==2.2.3
python-dotenv==1.0.1
